+++ halo2_gadgets/src/poseidon/primitives/p128pow5t3.rs
@@ -75,7 +75,7 @@ mod tests {
         super::{fp, fq},
         Fp, Fq,
     };
-    use crate::primitives::poseidon::{permute, ConstantLength, Hash, Spec};
+    use crate::poseidon::primitives::{permute, ConstantLength, Hash, Spec};
 
     /// The same Poseidon specification as poseidon::P128Pow5T3, but constructed
     /// such that its constants will be generated at runtime.
@@ -250,7 +250,7 @@ mod tests {
         {
             let (round_constants, mds, _) = super::P128Pow5T3::constants();
 
-            for tv in crate::primitives::poseidon::test_vectors::fp::permute() {
+            for tv in crate::poseidon::primitives::test_vectors::fp::permute() {
                 let mut state = [
                     Fp::from_repr(tv.initial_state[0]).unwrap(),
                     Fp::from_repr(tv.initial_state[1]).unwrap(),
@@ -268,7 +268,7 @@ mod tests {
         {
             let (round_constants, mds, _) = super::P128Pow5T3::constants();
 
-            for tv in crate::primitives::poseidon::test_vectors::fq::permute() {
+            for tv in crate::poseidon::primitives::test_vectors::fq::permute() {
                 let mut state = [
                     Fq::from_repr(tv.initial_state[0]).unwrap(),
                     Fq::from_repr(tv.initial_state[1]).unwrap(),
@@ -286,7 +286,7 @@ mod tests {
 
     #[test]
     fn hash_test_vectors() {
-        for tv in crate::primitives::poseidon::test_vectors::fp::hash() {
+        for tv in crate::poseidon::primitives::test_vectors::fp::hash() {
             let message = [
                 Fp::from_repr(tv.input[0]).unwrap(),
                 Fp::from_repr(tv.input[1]).unwrap(),
@@ -298,7 +298,7 @@ mod tests {
             assert_eq!(result.to_repr(), tv.output);
         }
 
-        for tv in crate::primitives::poseidon::test_vectors::fq::hash() {
+        for tv in crate::poseidon::primitives::test_vectors::fq::hash() {
             let message = [
                 Fq::from_repr(tv.input[0]).unwrap(),
                 Fq::from_repr(tv.input[1]).unwrap(),