+++ halo2_gadgets/src/primitives.rs
@@ -1,8 +1,0 @@
-//! Primitives used in the Orchard protocol.
-// TODO:
-// - DH stuff
-//     - EphemeralPublicKey
-//     - EphemeralSecretKey
-
-pub mod poseidon;
-pub mod sinsemilla;