+++ halo2_gadgets/src/sinsemilla/merkle.rs
@@ -318,7 +318,7 @@ pub mod tests {
                             (sibling, &node)
                         };
 
-                        use crate::primitives::sinsemilla;
+                        use crate::sinsemilla::primitives as sinsemilla;
                         let merkle_crh = sinsemilla::HashDomain::from_Q(TestHashDomain.Q().into());
 
                         merkle_crh